// src/generated/model/discountRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DiscountRequestDiscountType } from './discountRequestDiscountType.ts';
import type { DiscountRequestStatus } from './discountRequestStatus.ts';

export interface DiscountRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  code: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  name: string;
  discountType: DiscountRequestDiscountType;
  /** @minimum 0.01 */
  value: number;
  /** @minItems 1 */
  feeCategoryIds: string[];
  startDate: string;
  endDate: string;
  status: DiscountRequestStatus;
  description?: string;
}