// src/generated/model/updateDiscountRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateDiscountRequestDiscountType } from './updateDiscountRequestDiscountType.ts';
import type { UpdateDiscountRequestStatus } from './updateDiscountRequestStatus.ts';

export interface UpdateDiscountRequest {
  /**
     * @minLength 0
     * @maxLength 200
     */
  name: string;
  discountType: UpdateDiscountRequestDiscountType;
  /** @minimum 0.01 */
  value: number;
  /** @minItems 1 */
  feeCategoryIds: string[];
  startDate: string;
  endDate: string;
  status: UpdateDiscountRequestStatus;
  description?: string;
}