// src/generated/model/discountResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DiscountResponseDiscountType } from './discountResponseDiscountType.ts';
import type { DiscountResponseStatus } from './discountResponseStatus.ts';

export interface DiscountResponse {
  id?: string;
  code?: string;
  name?: string;
  discountType?: DiscountResponseDiscountType;
  value?: number;
  startDate?: string;
  endDate?: string;
  status?: DiscountResponseStatus;
  description?: string;
  feeCategoryIds?: string[];
  createdAt?: string;
  updatedAt?: string;
}